requirements.txt
# Core dependencies
ttkbootstrap>=1.10.0
requests>=2.28.0
psutil>=5.9.0

# File cracking dependencies
rarfile>=4.0
py7zr>=0.20.0
PyPDF2>=3.0.0
msoffcrypto-tool>=5.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
python-pptx>=0.6.21
xlrd>=2.0.1

# AI and ML dependencies
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
tqdm>=4.65.0

# GPU and cryptography
GPUtil>=1.4.0
pyopencl>=2022.3.1
pycryptodome>=3.15.0
cryptography>=3.4.8

# Image processing
Pillow>=9.0.0

# Build tools
pyinstaller>=5.0.0 